Hooks: rewrite the GdkPixbuf loader cache at runtime on Linux
The contents of GdkPixbuf's loader cache is only relocatable on
OSX/Windows or if a flag is specified at compile time
(see https://bugzilla.gnome.org/show_bug.cgi?id=737523). Since
we can't guarantee that it will be compiled with this flag,
instead we rewrite the cache at runtime to guarantee that it
will always point at the correct location, regardless of where
our app is.

Fixes #1833